Dynamically changing filters according to their filter type.
After a lot of sweat, swear and crying, we managed to get a way of getting the enabled filters and to change they color dynamically depending on the type of the filter. Moreover, some changes were made in the utils package (reorganizing and adding a new file) and some minor changes to reduce console errors and improve the working.
Hehe, keeping it up.