Connection to the API: succeeded
Managed to connect to the API, changing several classes. Adapter changed to connect with the API, added several types to  adapt it to the result and response format. Added filters to the bid cards and, despite not working properly, they are shown. Also, the data is properly displayed in the bid cards.